+++ webknossos/tests/dataset/test_dataset.py
@@ -1,5 +1,6 @@
 import itertools
 import json
+import os
 import pickle
 from pathlib import Path
 from typing import Iterator, Optional, Tuple, cast
@@ -16,10 +17,12 @@ from tests.constants import (
     TESTOUTPUT_DIR,
     use_minio,
 )
+from webknossos.client.context import webknossos_context
 from webknossos.dataset import (
     COLOR_CATEGORY,
     SEGMENTATION_CATEGORY,
     Dataset,
+    RemoteDataset,
     View,
 )
 from webknossos.dataset._array import DataFormat
@@ -41,11 +44,10 @@ from webknossos.utils import (
     snake_to_camel_case,
 )
 
-
-@pytest.fixture(autouse=True, scope="module")
-def start_minio() -> Iterator[None]:
-    with use_minio():
-        yield
+# @pytest.fixture(autouse=True, scope="module")
+# def start_minio() -> Iterator[None]:
+#     with use_minio():
+#         yield
 
 
 DATA_FORMATS = [DataFormat.WKW, DataFormat.Zarr]
@@ -1038,6 +1040,17 @@ def test_open_dataset_without_num_channels_in_properties() -> None:
         assert data["dataLayers"][0].get("numChannels") == 1
 
     assure_exported_properties(ds)
+
+
+@pytest.mark.use_proxay
+def test_explore_and_add_remote() -> None:
+    with webknossos_context("http://localhost:9000", os.environ.get("WK_TOKEN")):
+        remote_ds = RemoteDataset.explore_and_add_remote(
+            "http://localhost:9000/data/zarr/Organization_X/l4_sample/",
+            "added_remote_ds",
+            "/Datasets",
+        )
+        assert remote_ds.name == "added_remote_ds"
 
 
 def test_no_largest_segment_id() -> None: